+++ MQTT-Rev3.0/src/SSN_API/SSN_API.h
@@ -193,7 +193,7 @@ extern uint8_t SSN_CURRENT_SENSOR_RATINGS[4];
 /** SSN machine maximum loads for calculating percentage loads on machines */
 extern uint8_t SSN_CURRENT_SENSOR_MAXLOADS[4];
 /** SSN machine thresholds for deciding IDLE state */
-extern uint8_t SSN_CURRENT_SENSOR_THRESHOLDS[4];
+extern float SSN_CURRENT_SENSOR_THRESHOLDS[4];
 /** SSN Temperature and Humidity Sensor Thresholds */
 extern uint8_t TEMPERATURE_MIN_THRESHOLD, TEMPERATURE_MAX_THRESHOLD;
 extern uint8_t RELATIVE_HUMIDITY_MIN_THRESHOLD, RELATIVE_HUMIDITY_MAX_THRESHOLD;
